requirements/base.txt
django==2.2.4
psycopg2==2.8.3
django-crispy-forms==1.7.2
djangorestframework==3.10.2